Generic repository was created and registerd

+++ src/ArtExchange.Application/Feautures/Persons/Commands/CreatePersonCommand.cs
@@ -2,7 +2,7 @@
 
 namespace ArtExchange.Application.Feautures.Persons.Commands
 {
-    public class CreatePersonCommand:IRequest<int>
+    public class CreatePersonCommand:IRequest<long>
     {
         public string FirstName { get; set; } = string.Empty;
         public string LastName { get; set; } = string.Empty;
